refactor: 상품 가격 타입 Long에서 Integer로 변경

+++ product-service/src/main/java/com/destiny/productservice/presentation/dto/response/ProductResponse.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 public record ProductResponse(
     UUID id,
     String name,
-    Long price,
+    Integer price,
     UUID brandId,
     ProductStatus status,
     String color,
